Close terms of service when they are accepted
