+++ app/screens/terms_of_service/index.ts
@@ -5,6 +5,7 @@ import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
 import withObservables from '@nozbe/with-observables';
 
 import {observeConfigValue} from '@queries/servers/system';
+import {observeShowToS} from '@queries/servers/terms_of_service';
 
 import TermsOfService from './terms_of_service';
 
@@ -13,6 +14,7 @@ import type {WithDatabaseArgs} from '@typings/database/database';
 const enhanced = withObservables([], ({database}: WithDatabaseArgs) => {
     return {
         siteName: observeConfigValue(database, 'SiteName'),
+        showToS: observeShowToS(database),
     };
 });
